Migrate to Rust 2024
This raises the minimum supported Rust version (MSRV) to 1.85.0, uses
the rustfmt 2024 style edition, enables cargo's Rust-version aware
resolver (version 3), and refines captures and lifetimes (use bounds).

See <https://blog.rust-lang.org/2025/02/20/Rust-1.85.0/> for more
information.

+++ noodles-bcf/src/record/codec/value.rs
@@ -5,7 +5,7 @@ mod int32;
 mod int8;
 mod ty;
 
-pub use self::{array::Array, float::Float, int16::Int16, int32::Int32, int8::Int8, ty::Type};
+pub use self::{array::Array, float::Float, int8::Int8, int16::Int16, int32::Int32, ty::Type};
 
 #[derive(Debug)]
 pub enum Value<'a> {
